requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases
pydantic
passlib[bcrypt]
python-jose[cryptography]
qrcode
